fix(std.string): to_long 64-bit on Windows too — long long, not long
CI (Windows MinGW/MSYS2) surfaced that the prior commit only fixed half
the to_long truncation. Aether's `long` is 64-bit on every platform (it
lowers to int64_t), but the C parse path used C `long` + strtol — which
is 32-bit on Windows (LLP64), 64-bit only on LP64 Linux/macOS. So on
Windows string_to_long_raw wrote 4 bytes into Aether's 8-byte slot and
strtol overflowed on values > 2^31, still corrupting 64-bit values (and
breaking the string_to_long_raw raw extern the downstream relied on).

string_to_long_raw / string_get_long / string_try_long now use
`long long` + strtoll (matching string_from_long, which was already
long long). No-op on LP64 (long long == long there); correct on Windows.

The earlier commit's explicit `-> (long, string)` return type (the
inferred-multi-return long->int collapse) is still needed and unchanged;
this is the second, platform-specific half.

Verified: build; 226 C unit tests; regression test
(test_string_to_number_widths.ae) passes on Linux. The 64-bit assertions
(2^32, 48-bit, negative 2^33) now hold on Windows because strtoll no
longer overflows and the full 8 bytes round-trip.

+++ std/string/aether_string.h
@@ -187,7 +187,7 @@ AetherString* string_from_float(float value);
 // Aether code should prefer the Go-style wrappers in std.string
 // (`string.to_int`, etc.) which return `(value, error)` tuples.
 int string_to_int_raw(const void* str, int* out_value);
-int string_to_long_raw(const void* str, long* out_value);
+int string_to_long_raw(const void* str, long long* out_value);  // 64-bit slot (Windows `long` is 32-bit)
 int string_to_float_raw(const void* str, float* out_value);
 int string_to_double_raw(const void* str, double* out_value);
 
@@ -196,8 +196,8 @@ int string_to_double_raw(const void* str, double* out_value);
 // zero-value on failure.
 int    string_try_int(const void* s);
 int    string_get_int(const void* s);
-int    string_try_long(const void* s);
-long   string_get_long(const void* s);
+int       string_try_long(const void* s);
+long long string_get_long(const void* s);  // 64-bit on every platform
 int    string_try_float(const void* s);
 double string_get_float(const void* s);  // 64-bit: Aether `float` lowers to C double
 int    string_try_double(const void* s);
